move content-specific tests to spec_suites/content/
