+++ test/browser_tests/spec_suites/content/about-director.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var TheAboutDirectorPage = require( '../page_objects/page_about-director.js' );
+var TheAboutDirectorPage = require( '../../page_objects/page_about-director.js' );
 
 describe( 'The About Director Page', function() {
   var page;
