+++ test/browser_tests/spec_suites/content/about-us.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var AboutUs = require(
-    '../page_objects/page_about-us.js'
+    '../../page_objects/page_about-us.js'
   );
 
 describe( 'About Landing Page', function() {
