+++ test/browser_tests/spec_suites/content/dbwu-small-businesses.js
@@ -1,9 +1,9 @@
 'use strict';
 
 var SmallBusinessess =
-require( '../page_objects/page_dbwu-small-businesses.js' );
+require( '../../page_objects/page_dbwu-small-businesses.js' );
 
-var Urlmatcher = require( '../util/url-matcher' );
+var Urlmatcher = require( '../../util/url-matcher' );
 
 describe( 'The Small Businesses Page', function() {
   var page;
