+++ test/browser_tests/spec_suites/content/homepage.js
